refactor: move debugger from core internal to default integration
Debugging depends on nvim-dap, an optional third-party plugin, putting it
in the same category as the luasnip/treesitter integrations rather than the
always-on core modules (lsp, buf, workspace, config, multiproject).

As a core internal it had two problems: the top-level require("dap") in
debugger.lua was silently swallowed by the pcall in config.lua for users
without nvim-dap, and the scheduled debugger.setup() could run before a
lazy-loaded dap was available. The integration loader already solves both
via its LazyLoad autocmd.

- Fold the maintained debugger.lua body into integrations/dap.lua (the old
  copy there was a stale duplicate), deferring the dap require into setup()
  and pcall-guarding nvim-dap-virtual-text
- Delete lua/al/debugger.lua
- Enable dap by default in config integrations; drop the pcall(debugger.setup)
- Fix p2i key "dap" -> "nvim-dap" so the LazyLoad autocmd actually matches
- Drop the al.debugger nvimSkipModule entry (no top-level plugin require left)
