+++ lua/al/integrations/init.lua
@@ -6,7 +6,7 @@ local M = {}
 M.p2i = {
     ["LuaSnip"] = "luasnip",
     ["nvim-treesitter"] = "treesitter",
-    ["dap"] = "dap",
+    ["nvim-dap"] = "dap",
 }
 
 -- integration to plugin
